Backend/Data service/VectorDB/requirements.txt
faiss-cpu
sentence-transformers
llama-cpp-python
langchain
langchain-community
gpt4all
beautifulsoup4
unstructured
torch
pypdf
openai
ctransformers
transformers
docx
python-docx
bs4
html2text
pymupdf
requests

#